requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
form==0.0.1
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
mysqlclient==2.0.3
packaging==20.9
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
xhtml2pdf==0.2.5
